Split off auto_quit tests for quicker integration runs
